Fixes language switching to default language
Closes #2632

From this commit on the language id won't be stored in the session any more
for frontend requests.

Upon every page visit the language id got stored in the session. When switching from a secondary language to the default language no language param is passed to the controller. The order how the language is found prioritized the session and the default was just a fallback. Coming from a secondary language and therefore it's id in the session, it would never load the default language, but the secondary.

Storing the language in the session is still done in the admin section.