Set admin picture thumbnail quality to 90
Make the picture thumbnail quality independent from the
frontend output_image_quality config.

Using 90 as value for quality should be a good compromise
between file size and image quality.

The same image with quality of 100 is three times the size of an image with quality of 90 while having nearly the same image
quality for a small thumbnail. An large thumbnail (as used in the
image cropper) is ten times the size in 100% quality!