Implement custom scrubber for Alchemy::Ingredients::Richtext
We found that using Rails' HTML sanitizer does more than we want the
Richtext sanitization to do: It does not just remove nodes that are not
in the safelist, it also escapes some markup (especially in links).

This introduces a custom Loofah "scrubber" that only cares about the
element safelist.

The `sanitized_body` attribute is not for escaping at the view layer,
where all these safety precautions are necessary, but just for making
sure admin's don't use iframes when we don't want to.

See the following related issues and commits:
https://github.com/rails/rails-html-sanitizer/commit/f3ba1a839a35f2ba7f941c15e239a1cb379d56ae
https://github.com/sparklemotion/nokogiri/issues/3104
https://github.com/sparklemotion/nokogiri/issues/969#issuecomment-973068692
https://github.com/flavorjones/loofah/issues/14#issuecomment-1006617141