Fix ActionNotFound in PictureController
The default setting of Rails 7.1 will raise an exception if a callback action was not found. The info - method is not available in picture controller and will lead to an exception if the option is turned on in Rails 7.1.

(cherry picked from commit aa467525da1df94ea1ecd00706e84119459936b5)