Toggle sortable mode on sitemap
In order to prevent accidental moves (that would alter the page
url) users now need to enable sortable mode.

We hide the tooltip on the page icon (which is the draggable
handle) in order to make it easier to see that move cursor for
locked pages.

Coloring the background light yellow indicates that we now
are in a special mode.