Locally initialize SortableElements
We have two places where we need to initialize SortableJS
for elements.

1) The ElementsWindow and its top level elements
2) The nested elements inside of each element-editor

Instead of using the global Alchemy object to initialize
them we can import the file locally and init it where it needs
to be. This makes it possible to remove another global
call from the `create.js.erb` file.