Use page version's `updated_at` timestamp as cache key
Currently, it's hard for us to invalidate caches for Alchemy pages when
content that's referenced through ingredients with related objects
changes.

For example, in a situation where a user combines Alchemy and Solidus
using the `alchemy_solidus` gem, a page's cache key does not update when
a product that's referenced through a SpreeProduct ingredient changes.

There's a PR up on the alchemy_solidus gem that touches ingredients in
these situations, and with this change, that touching can be used for
breaking caches. [1]

In the unlikely event the requested version is not available, we use the
page's updated_at timestamp as a fallback.

The core of the logic has been moved to a new `#last_modified_at`
method that can be used for the last-modified header in HTTP responses.
[1](https://github.com/AlchemyCMS/alchemy-solidus/pull/108)