Replace ingredient update - js.erb with turbo stream
As a preparation to replace the Rails Javascript responses, the ingredient update response was replaced with turbo stream. It now uses the alchemy-action - component to trigger action in our frontend. As an addition the response also triggers a previewReload action to reflect the changes in the settings of an ingredient.