Add Language and Site serializers
Without those ActiveModelSerializers can't find a serializer
for the site and language relations in the PageSerializer which
is used in the PageSelect. It delegates serialization to Rails
which takes whatever serializer is responsible for to_json,
which in some cases lead to nesting the site or language with
it's own root key, which is not what we want.