Disable Turbo Cache in Admin
This feature does not provide a huge benefit and has some weird downsides. If the user is revisiting a view in Alchemy admin it will be slower, but at least the user is not seeing and old view which will updating a few moments later (e.g. already closed tabs, newly created pictures, etc.)