Fix combining search filters and pagination
The `search_filter_params` method in Alchemy's Resource Controller
converts the permitted attributes to a Hash, stopping them from being
recognized as something like a nested Hash in our per page select. This
commit uses duck-typing in order to find out whether we're looking at a
nested Hash. This should work with both ActionController::Parameters and
with a Hash.