Extract Attribute into a cell and header - component
Use more view components to simplify the table and make it more readable. This also improves the before_render - method in the table. The content - "hack" is now not necessary anymore, because the view_context isn't used anymore and all other view components are placed as slots in the table.