Fallback to @page var if no Current.page is set
With Alchemy 7.2 (#2701) we started to use `Alchemy::Current.page`
for rendering elements of current page. If one has a custom
controller to render Alchemy content with:

    render(template: "alchemy/pages/show")

it is now empty if `Alchemy::Current.page` is not set.

Someday we might deprecate the `@page` instance variable,
but until then we need to fall back to it.