Use alchemy_display_name for page actor names
If you have a user class that does not implement
name (eg. `Spree::User` or a generic `User` class)
it would return `nil` for Page actor methods (creator,
updater, locker).

We already use `alchemy_display_name` for the current
logged in user displayed in the admin frame.

Using this for the Page actor methods to be consistant
and allow custom user classes to present a user by it.