Load jquery via importmap
We still need it for select2 but we want to update and install it via
node modules.

In order for this to work we need to use module script tags. Since
jQuery is loaded via importmap and a JS module (`alchemy_admin.js`)
it execution is deferred.

(cherry picked from commit 770a0d0cf524bb203cbc276e4e96a295e7f61970)