Add support for Propshaft
Adds Propshaft support but still maintain Sprockets support.

In order to make this happen we need to deal with Tinymce
assets. They need to be available without the digest. For Sprockets
we are using the `NonStupidDigestAssets` gem.

For Propshaft we enhance the `assets:precompile` Rake task and
simply copy the digested assets to the non digested name.
In development we return the undigested asset if requested.