Use turbo streams to update page from configure dialog
This one was complex because we do a lot of different things
dependent of where we are and what page type we update.

The sitemap tree still needs to be replaced by calling a
script, due to the fact that it is rendered with Handlerbars.

Luckily Turbo let's us do this as well and can tackle that problem
later.