Generate a view component in ingredient generator
Alchemy is mostly using a ViewComponent to render an ingredient. Only some situations the view partial is used. If you previous used the generator you ran into the situation, that the required ingredient was missing. Now the generator is creating the ingredient view component and is referring to that component in the view.

(cherry picked from commit 6bb2595795d508c915935ad1cf37eb0c54c04423)