Fix loading custom properties into Tinymce skin
For the Tinymce skin we need our custom properties available.
We cannot simply compile them into the skin, because it gets
loaded from Tinymce into the current document and this would
lead to duplicated custom properties.

So instead we compile it as separate file and include it in the
Alchemy admin as first stylesheet. This has the additional advantage
that we can load this file into ie. Solidus to display the Alchemy
Tinymce skin as well.