Persist Elements Window width in session
Makes sure the elements window width is the same across pages.
Uses a cookie, so we can read the value from Rails and put it on
the body tag, so it is immediatly available on page render.
Otherwise we would see the element window resizing on every
page visit. SSR ftw \o/