Add link to installation guide
The guides already cover the relevant dependencies
needed to run AlchemyCMS.

As for contributing to AlchemyCMS or changing it,
we don't currently have a dedicated place to list
additional dependencies like chromedriver.

I think the README would be a good spot to include
that information, but it could also make its way
into the guides if that makes more sense. What do
you think? Shall we add a section on additional
dependencies to the README or the guides?