Deprecate Alchemy::Config for Alchemy.config
The `const_missing` needs to be defined both as a class method and as an
instance method in order to catch references to Alchemy::Config from the
class level and from the instance level. To avoid repetition, I've moved
the method into its own module that we now both `include` and `extend`
in the main `Alchemy` module.