Use Ransack for Filtering
This removes our custom filter implementation, and moves it to be
ransack-only.

The existing features we had are comfortably implementable using
ransacks `:q` parameter, and this does it across the app. The code used
for searching now lives in the `Alchemy::Admin::ResourceFilters` module.

This also removes the legacy "filter" parameters from the search form.