Add a CollectionOption class
We have a few configurations that are currently arrays of Hashes. This
is not type-safe, and sometimes hard to use. This adds a
`Alchemy::Configuration::ConfigurationOption` class that can be used
with the `CollectionOption` class in `Alchemy::Configuration`. It should not be
used as a plain option (we have the `configuration` method for that).

(cherry picked from commit 555091c8897c0180cfe81542dae3b0b84708321b)