Translate include blank per filter
Select filters start out with a string that currently always says "All".
This commit allows developers to set a nice string that indicates the
absence of the filter. If the string is not set, we'll still see "All",
which makes sense in many cases.

The previous translation has not been used once I believe.