Copy new config initializer during upgrade
We change our configuration to be an ruby file instead of
the `config/alchemy/config.yml` file. This yml file will still
be read from, but we encourage people to update their
config.

This changes how we setup the upgrader. We now instantiate
the upgrader instance in the rake task and include each version
as modules instead of individual classes. That way we can
make use of Thor actions much easier than before.