Pass picture to Picture.url_class
Moves the PictureVariant usage into the Url class.
The variant is only used for the dragonfly url, so
we move this implementation detail into this namespace.

This removes the debatable feature to pass additional
params to the picture url. A feature that never made any sense
and will not be supported in ActiveStorage.