Use image_file_extension for picture variant rendering
In ActiveStorage the `image_file_format` will be the file
mime type and not the file extension. Since Dragonfly
uses the file extension as format in it's "magic column"
we introduced the `image_file_extension` method in a
previous commit, but now need to prepare the picture
variant for the upcoming adapter.