fix(alchemy-growl): Remove after inserting to DOM
We attach the `<alchemy-growl>` custom element to the
DOM with turbo stream responses, but we never want
to display the growl message in the body. It calls
the `growl()` function which adds a toast message to
the top right corner.

(cherry picked from commit 8355dff4584820e5f9f236c098be6b8c5b7e9623)