Resources Admin: Fix editing with filters
Without this change, admins would lose all but the first of their search
filters on editing a resource, because the `&` between subsequent
filters would be escaped to `&amp;`, making the values appear in the
params hash as `params[amp;q]`. This parameter would not be recognized
as search parameter and discarded.