Resources Admin: Display search string in search field
When navigating around the resources admin for e.g. events, and having
searched in the standard search field, it's nice to have that search
field's value prefilled with whatever has been searched for before so
that one does not lose the search term on the next navigation.

This also changes the spec to use "Karaoke" instead of "Show" because
there's a lot of "Show" in a CRUD admin.