fix(picture_spec): Ensure picture is created
If we do not ensure that the image is created before attaching it
to the ingredient we run into race conditions leading to "IOError: closed stream"
errors with active storage. This is the more realistic scenario anyway.

(cherry picked from commit 125e0a0ed995c27f4099515657cb13c9216c83e5)