Use .gitattributes to ignore unwanted files in gem builds
Git has a feature to specify unwanted files for builds, the
`gitattributes` file with the `export-ignore` directive. This uses that
in order to construct the list of files for the gem. This way the
`spec/` directory with its non-standard filenames won't appear anywhere
in the pipeline.

Also removes a regular expression from the codebase, which is always
good.