Nullify related_object_type if related_object_id is set to nil
If a related object is removed from in ingredient the id gets nullified
by the form field. We need to make sure that the related_object_type
is also set to nil, because it is an polymorphic relation.

(cherry picked from commit 5657704f4b663fe140f8613a5623c7c6608379cb)