Allow stale-while-revalidate cache-control header
If you have a CDN that supports the `stale-while-revalidate` header
(like Cloudflare [^1]) - you can enable this header for better CDN
performance.

Set this value pages in the new

    Alchemy.config.page_cache.stale_while_revalidate

configuration option to any integer val.ue in seconds.

It is disabled by default, because this only works with CDNs. Browsers will
ignore this caching rule and we use `must-revalidate` as rule by default.

[^1]: https://developers.cloudflare.com/cache/concepts/cache-control/#revalidation