Remove three letter urlname restriction
This was implemented to make sure the root url language code
requests are not overwritten by any page requested with short
urlnames. This is handled by the router already and only leads
to problems on nested urls, where a page with short names get
urlnames with dashes in the url. In the age of AI this might
not what we want. We want `example.com/services/ai` instead
of `example.com/services/-ai` ;)