Add color ingredient type
Add a new ingredient type to select a color. By default only a color input will be shown. It is possible to provide predefined colors as settings and then the color ingredient will show a select of all colors. Furthermore it has a second option 'custom_color' which adds the possibility to select a color which is not predefined, if necessary.