Fix showing picture detail in overlay
We lost the picture_counter during a previous refactor.
But we need it for loading the correct image in the
detail overlay. In that form we also need to pass the
page param dependent from the previous or next
image counter.

(cherry picked from commit 32c51d36fbe54cab6a2995e2c1d889f5365d58e0)