Fix `stub_alchemy_config` (and make it recursive)
This allows us to deeply stub the Alchemy configuration using nested
hashes as an argument. Also, stubbing `Alchemy.config` will not reset
the config object any longer, as it did before. This has especially been
a problem for things like the importmap configuration, as that would be
added on engine load, but would not be a "default" setting, as it is
with many of the scalar values in the configuration.