Remove config.auth.user_class override
The purpose of this override was to provide a friendly error message in
case a developer has not configured a user class. This now happens with
a custom error in Alchemy's `before_initialize` block (so before the
Autoloader kicks in). This removes a lot of complexity from the auth
class.