New Update Checks services
Instead of running update checks always to Rubygems
we now perform the check by default against a new
app.alchemy-cms.com service.

The current Alchemy version, Rails version, Ruby version
and the host Alchemy is running on will be shared with the
maintainers of AlchemyCMS (Blish GmbH). No PII is transmitted
and the data is soley used for analytic purposes and securely
stored.

If you do not want that, you can set it back to rubygems.org
by setting

    Alchemy.config.update_check_service = :ruby_gems

or even completely disable the update check by setting

    Alchemy.config.update_check_service = :none

Signed-off-by: Thomas von Deyen <vondeyen@blish.cloud>